Add Jansson stub for cleaner non-Jansson compilations
